Extract emulator mouse and cursor helpers
